Seperated the utils
